Add TestNG suite and DynamoDbLocal test
Adds a TestNG suite so that suite-level commands (such as starting a
database) work. Wires that into the Gradle build.

Gets the first test of DynamoDb local working by modifying the model to
fit some assumptions the DDB SDK makes (such as getter methods strictly
following the "getX()" syntax as opposed to the newer record style of
"x()".